Extract discrepancy logic
