+++ app/view_models/effort_audit_view.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
 class EffortAuditView < EffortWithLapSplitRows
-
-  DISCREPANCY_THRESHOLD = 1.minute
-
   delegate :event_name, :person, :start_time, :has_start_time?, :stopped?, to: :loaded_effort
   delegate :simple?, :multiple_sub_splits?, :laps_unlimited?, :event_group, to: :event
